Moving func to ChartRenderHints
