+++ no.marintek.mylyn.wikitext.ooxml.core/src/no/marintek/mylyn/wikitext/ooxml/ChartRenderHints.java
@@ -19,6 +19,10 @@ public class ChartRenderHints {
 	private int xAxisNumberFormat;
 	
 	private int yAxisNumberFormat;
+	
+	private LineStyle[] lineStyles;
+	
+	private PointStyle[] pointStyles;
 
 	public static int AUTOFORMAT = 0;
 	public static int GENERAL = 1;
@@ -86,5 +90,28 @@ public class ChartRenderHints {
 	public void setRichLegend(boolean richLegend) {
 		this.richLegend = richLegend;
 	}
+	
+	public LineStyle[] getLineStyles() {
+		return lineStyles;
+	}
+	
+	/**
+	 * Set the LineStyle for each Plot in the series
+	 * @param lineStyles
+	 */
+	public void setLineStyles(LineStyle[] lineStyles) {
+		this.lineStyles = lineStyles;
+	}
 
+	public PointStyle[] getPointStyles() {
+		return pointStyles;
+	}
+	
+	/**
+	 * Set the PointStyle for each Plot in the series
+	 * @param pointStyles
+	 */
+	public void setPointStyles(PointStyle[] pointStyles) {
+		this.pointStyles = pointStyles;
+	}
 }
