--- no.marintek.mylyn.wikitext.ooxml.core/src/no/marintek/mylyn/wikitext/ooxml/PlotStyleHints.java
@@ -1,23 +1,0 @@
-package no.marintek.mylyn.wikitext.ooxml;
-
-public class PlotStyleHints {
-	
-	private LineStyle[] lineStyles;
-	private PointStyle[] pointStyles;
-
-	public LineStyle[] getLineStyles() {
-		return lineStyles;
-	}
-	
-	public void setLineStyles(LineStyle[] lineStyles) {
-		this.lineStyles = lineStyles;
-	}
-
-	public PointStyle[] getPointStyles() {
-		return pointStyles;
-	}
-	
-	public void setPointStyles(PointStyle[] pointStyles) {
-		this.pointStyles = pointStyles;
-	}
-}
